separate market updates from snapshot
