+++ subgraphs/compound-v3/src/sdk/snapshots.ts
@@ -339,7 +339,7 @@ export class SnapshotManager {
   ///// Updaters /////
   ////////////////////
 
-  public updateUsageData(transactionType: string, account: Bytes): void {
+  updateUsageData(transactionType: string, account: Bytes): void {
     const dailyActiveAccountID = AccountActivity.DAILY.concat("-")
       .concat(account.toHexString())
       .concat("-")
@@ -351,10 +351,6 @@ export class SnapshotManager {
       .concat(account.toHexString())
       .concat("-")
       .concat(this.marketHourlySnapshot.hours.toString());
-    const activeAccountMarketID = account
-      .toHexString()
-      .concat("-")
-      .concat(this.market.id.toHexString());
     const activeAccountTransactionID = transactionType
       .concat("-")
       .concat(account.toHexString());
@@ -372,16 +368,10 @@ export class SnapshotManager {
       dailyActiveAccountMarketID
     ); // market daily
     let hourlyActiveAccount = _ActiveAccount.load(hourlyActiveAccountID); // usage hourly
-    let activeAccountMarket = _ActiveAccount.load(activeAccountMarketID); // market
-    let activeAccountTransaction = _ActiveAccount.load(
-      activeAccountTransactionID
-    ); // lending protocol
     let dailyActiveAccountTransaction = _ActiveAccount.load(
       dailyActiveAccountTransactionID
     ); // usage daily
-    let activeAccountTransactionMarket = _ActiveAccount.load(
-      activeAccountTransactionMarketID
-    ); // market
+
     let dailyActiveAccountTransactionMarket = _ActiveAccount.load(
       dailyActiveAccountTransactionMarketID
     ); // market daily
@@ -404,23 +394,6 @@ export class SnapshotManager {
       hourlyActiveAccount.save();
       this.usageHourlySnapshot.hourlyActiveUsers += INT_ONE;
     }
-    if (!activeAccountMarket && transactionType != TransactionType.LIQUIDATEE) {
-      activeAccountMarket = new _ActiveAccount(activeAccountMarketID);
-      activeAccountMarket.save();
-      this.market.cumulativeUniqueUsers += INT_ONE;
-    }
-    if (!activeAccountTransaction) {
-      activeAccountTransaction = new _ActiveAccount(activeAccountTransactionID);
-      activeAccountTransaction.save();
-      if (transactionType == TransactionType.DEPOSIT)
-        this.protocol.cumulativeUniqueDepositors += INT_ONE;
-      if (transactionType == TransactionType.BORROW)
-        this.protocol.cumulativeUniqueBorrowers += INT_ONE;
-      if (transactionType == TransactionType.LIQUIDATOR)
-        this.protocol.cumulativeUniqueLiquidators += INT_ONE;
-      if (transactionType == TransactionType.LIQUIDATEE)
-        this.protocol.cumulativeUniqueLiquidatees += INT_ONE;
-    }
     if (!dailyActiveAccountTransaction) {
       dailyActiveAccountTransaction = new _ActiveAccount(
         dailyActiveAccountTransactionID
@@ -434,24 +407,6 @@ export class SnapshotManager {
         this.usageDailySnapshot.dailyActiveLiquidators += INT_ONE;
       if (transactionType == TransactionType.LIQUIDATEE)
         this.usageDailySnapshot.dailyActiveLiquidatees += INT_ONE;
-    }
-    if (!activeAccountTransactionMarket) {
-      activeAccountTransactionMarket = new _ActiveAccount(
-        activeAccountTransactionMarketID
-      );
-      activeAccountTransactionMarket.save();
-      if (transactionType == TransactionType.DEPOSIT)
-        this.market.cumulativeUniqueDepositors += INT_ONE;
-      if (transactionType == TransactionType.BORROW)
-        this.market.cumulativeUniqueBorrowers += INT_ONE;
-      if (transactionType == TransactionType.LIQUIDATOR)
-        this.market.cumulativeUniqueLiquidators += INT_ONE;
-      if (transactionType == TransactionType.LIQUIDATEE)
-        this.market.cumulativeUniqueLiquidatees += INT_ONE;
-      if (transactionType == TransactionType.TRANSFER)
-        this.market.cumulativeUniqueTransferrers += INT_ONE;
-      if (transactionType == TransactionType.FLASHLOAN)
-        this.market.cumulativeUniqueFlashloaners += INT_ONE;
     }
     if (!dailyActiveAccountTransactionMarket) {
       dailyActiveAccountTransactionMarket = new _ActiveAccount(
@@ -472,11 +427,9 @@ export class SnapshotManager {
         this.marketDailySnapshot.dailyActiveFlashloaners += INT_ONE;
     }
 
-    this.market.save();
-    this.protocol.save();
     this.createOrUpdateUsageDailySnapshot();
     this.createOrUpdateUsageHourlySnapshot();
-    this.marketDailySnapshot.save();
+    this.createOrUpdateMarketDailySnapshot();
   }
 
   updateTransactionData(
@@ -485,12 +438,6 @@ export class SnapshotManager {
     amountUSD: BigDecimal
   ): void {
     if (transactionType == TransactionType.DEPOSIT) {
-      this.protocol.depositCount += INT_ONE;
-      this.protocol.cumulativeDepositUSD =
-        this.protocol.cumulativeDepositUSD.plus(amountUSD);
-      this.market.cumulativeDepositUSD =
-        this.market.cumulativeDepositUSD.plus(amountUSD);
-      this.market.depositCount += INT_ONE;
       this.marketDailySnapshot.dailyDepositUSD =
         this.marketDailySnapshot.dailyDepositUSD.plus(amountUSD);
       this.marketDailySnapshot.dailyNativeDeposit =
@@ -502,8 +449,6 @@ export class SnapshotManager {
       this.usageDailySnapshot.dailyDepositCount += INT_ONE;
       this.usageHourlySnapshot.hourlyDepositCount += INT_ONE;
     } else if (transactionType == TransactionType.WITHDRAW) {
-      this.protocol.withdrawCount += INT_ONE;
-      this.market.withdrawCount += INT_ONE;
       this.marketDailySnapshot.dailyWithdrawUSD =
         this.marketDailySnapshot.dailyWithdrawUSD.plus(amountUSD);
       this.marketDailySnapshot.dailyNativeWithdraw =
@@ -515,12 +460,6 @@ export class SnapshotManager {
       this.usageDailySnapshot.dailyWithdrawCount += INT_ONE;
       this.usageHourlySnapshot.hourlyWithdrawCount += INT_ONE;
     } else if (transactionType == TransactionType.BORROW) {
-      this.protocol.borrowCount += INT_ONE;
-      this.protocol.cumulativeBorrowUSD =
-        this.protocol.cumulativeBorrowUSD.plus(amountUSD);
-      this.market.cumulativeBorrowUSD =
-        this.market.cumulativeBorrowUSD.plus(amountUSD);
-      this.market.borrowCount += INT_ONE;
       this.marketDailySnapshot.dailyBorrowUSD =
         this.marketDailySnapshot.dailyBorrowUSD.plus(amountUSD);
       this.marketDailySnapshot.dailyNativeBorrow =
@@ -532,8 +471,6 @@ export class SnapshotManager {
       this.usageDailySnapshot.dailyBorrowCount += INT_ONE;
       this.usageHourlySnapshot.hourlyBorrowCount += INT_ONE;
     } else if (transactionType == TransactionType.REPAY) {
-      this.protocol.repayCount += INT_ONE;
-      this.market.repayCount += INT_ONE;
       this.marketDailySnapshot.dailyRepayUSD =
         this.marketDailySnapshot.dailyRepayUSD.plus(amountUSD);
       this.marketDailySnapshot.dailyNativeRepay =
@@ -545,12 +482,6 @@ export class SnapshotManager {
       this.usageDailySnapshot.dailyRepayCount += INT_ONE;
       this.usageHourlySnapshot.hourlyRepayCount += INT_ONE;
     } else if (transactionType == TransactionType.LIQUIDATE) {
-      this.protocol.liquidationCount += INT_ONE;
-      this.protocol.cumulativeLiquidateUSD =
-        this.protocol.cumulativeLiquidateUSD.plus(amountUSD);
-      this.market.cumulativeLiquidateUSD =
-        this.market.cumulativeLiquidateUSD.plus(amountUSD);
-      this.market.liquidationCount += INT_ONE;
       this.marketDailySnapshot.dailyLiquidateUSD =
         this.marketDailySnapshot.dailyLiquidateUSD.plus(amountUSD);
       this.marketDailySnapshot.dailyNativeLiquidate =
@@ -562,10 +493,6 @@ export class SnapshotManager {
       this.usageDailySnapshot.dailyLiquidateCount += INT_ONE;
       this.usageHourlySnapshot.hourlyLiquidateCount += INT_ONE;
     } else if (transactionType == TransactionType.TRANSFER) {
-      this.protocol.transferCount += INT_ONE;
-      this.market.cumulativeTransferUSD =
-        this.market.cumulativeTransferUSD.plus(amountUSD);
-      this.market.transferCount += INT_ONE;
       this.marketDailySnapshot.dailyTransferUSD =
         this.marketDailySnapshot.dailyTransferUSD.plus(amountUSD);
       this.marketDailySnapshot.dailyNativeTransfer =
@@ -576,10 +503,6 @@ export class SnapshotManager {
         this.financialSnapshot.dailyTransferUSD.plus(amountUSD);
       this.usageDailySnapshot.dailyTransferCount += INT_ONE;
     } else if (transactionType == TransactionType.FLASHLOAN) {
-      this.protocol.flashloanCount += INT_ONE;
-      this.market.cumulativeFlashloanUSD =
-        this.market.cumulativeFlashloanUSD.plus(amountUSD);
-      this.market.flashloanCount += INT_ONE;
       this.marketDailySnapshot.dailyFlashloanUSD =
         this.marketDailySnapshot.dailyFlashloanUSD.plus(amountUSD);
       this.marketDailySnapshot.dailyNativeFlashloan =
@@ -595,18 +518,14 @@ export class SnapshotManager {
       ]);
       return;
     }
-    this.protocol.transactionCount += INT_ONE;
-    this.market.transactionCount += INT_ONE;
     this.usageDailySnapshot.dailyTransactionCount += INT_ONE;
     this.usageHourlySnapshot.hourlyTransactionCount += INT_ONE;
 
-    this.protocol.save();
-    this.market.save();
     this.createOrUpdateMarketDailySnapshot();
     this.createOrUpdateMarketHourlySnapshot();
     this.createOrUpdateFinancials();
-    this.usageDailySnapshot.save();
-    this.usageHourlySnapshot.save();
+    this.createOrUpdateUsageDailySnapshot();
+    this.createOrUpdateUsageHourlySnapshot();
   }
 
   updateRevenue(
